refactor: flyttet kode til ny komponent
